+++ msp/msp_test.go
@@ -11,7 +11,6 @@ import (
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
-	"crypto/sha256"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/asn1"
@@ -20,6 +19,7 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
+	"net"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -33,7 +33,6 @@ import (
 	"github.com/hyperledger/fabric/bccsp/sw"
 	"github.com/hyperledger/fabric/bccsp/utils"
 	"github.com/hyperledger/fabric/core/config/configtest"
-	"github.com/hyperledger/fabric/protoutil"
 	"github.com/stretchr/testify/require"
 )
 
@@ -285,162 +284,6 @@ func TestSerializeIdentities(t *testing.T) {
 	}
 }
 
-func computeSKI(key *ecdsa.PublicKey) []byte {
-	raw := elliptic.Marshal(key.Curve, key.X, key.Y)
-	hash := sha256.Sum256(raw)
-	return hash[:]
-}
-
-func TestValidHostname(t *testing.T) {
-	tests := []struct {
-		name  string
-		valid bool
-	}{
-		{"", false},
-		{".", false},
-		{"example.com", true},
-		{"example.com.", true},
-		{"*.example.com", true},
-		{".example.com", false},
-		{"host.*.example.com", false},
-		{"localhost", true},
-		{"-localhost", false},
-		{"Not_Quite.example.com", true},
-		{"weird:colon.example.com", true},
-		{"1-2-3.example.com", true},
-	}
-	for _, tt := range tests {
-		if tt.valid {
-			require.True(t, validHostname(tt.name), "expected %s to be a valid hostname", tt.name)
-		} else {
-			require.False(t, validHostname(tt.name), "expected %s to be an invalid hostname", tt.name)
-		}
-	}
-}
-
-func TestValidateCANameConstraintsMitigation(t *testing.T) {
-	// Prior to Go 1.15, if a signing certificate contains a name constraint, the
-	// leaf certificate does not include a SAN, and the leaf common name looks
-	// like a valid hostname, the certificate chain would fail to validate.
-	// (This behavior may have been introduced with Go 1.10.)
-	//
-	// In Go 1.15, the behavior has changed and, by default, the same structure
-	// will validate. This test asserts on the old behavior.
-	caKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	require.NoError(t, err)
-	leafKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	require.NoError(t, err)
-
-	caKeyUsage := x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment | x509.KeyUsageCertSign | x509.KeyUsageCRLSign
-	caTemplate := x509.Certificate{
-		Subject:                     pkix.Name{CommonName: "TestCA"},
-		SerialNumber:                big.NewInt(1),
-		NotBefore:                   time.Now().Add(-1 * time.Hour),
-		NotAfter:                    time.Now().Add(2 * time.Hour),
-		ExcludedDNSDomains:          []string{"example.com"},
-		PermittedDNSDomainsCritical: true,
-		IsCA:                        true,
-		BasicConstraintsValid:       true,
-		KeyUsage:                    caKeyUsage,
-		SubjectKeyId:                computeSKI(caKey.Public().(*ecdsa.PublicKey)),
-	}
-	caCertBytes, err := x509.CreateCertificate(rand.Reader, &caTemplate, &caTemplate, caKey.Public(), caKey)
-	require.NoError(t, err)
-	ca, err := x509.ParseCertificate(caCertBytes)
-	require.NoError(t, err)
-
-	leafTemplate := x509.Certificate{
-		Subject:      pkix.Name{CommonName: "localhost"},
-		SerialNumber: big.NewInt(2),
-		NotBefore:    time.Now().Add(-1 * time.Hour),
-		NotAfter:     time.Now().Add(2 * time.Hour),
-		KeyUsage:     x509.KeyUsageDigitalSignature,
-		SubjectKeyId: computeSKI(leafKey.Public().(*ecdsa.PublicKey)),
-	}
-	leafCertBytes, err := x509.CreateCertificate(rand.Reader, &leafTemplate, ca, leafKey.Public(), caKey)
-	require.NoError(t, err)
-
-	keyBytes, err := x509.MarshalPKCS8PrivateKey(leafKey)
-	require.NoError(t, err)
-
-	caCertPem := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: caCertBytes})
-	leafCertPem := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: leafCertBytes})
-	leafKeyPem := pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: keyBytes})
-
-	t.Run("VerifyNameConstraintsSingleCert", func(t *testing.T) {
-		for _, der := range [][]byte{caCertBytes, leafCertBytes} {
-			cert, err := x509.ParseCertificate(der)
-			require.NoError(t, err, "failed to parse certificate")
-
-			err = verifyLegacyNameConstraints([]*x509.Certificate{cert})
-			require.NoError(t, err, "single certificate should not trigger legacy constraints")
-		}
-	})
-
-	t.Run("VerifyNameConstraints", func(t *testing.T) {
-		var certs []*x509.Certificate
-		for _, der := range [][]byte{leafCertBytes, caCertBytes} {
-			cert, err := x509.ParseCertificate(der)
-			require.NoError(t, err, "failed to parse certificate")
-			certs = append(certs, cert)
-		}
-
-		err = verifyLegacyNameConstraints(certs)
-		require.Error(t, err, "certificate chain should trigger legacy constraints")
-		var cie x509.CertificateInvalidError
-		require.True(t, errors.As(err, &cie))
-		require.Equal(t, x509.NameConstraintsWithoutSANs, cie.Reason)
-	})
-
-	t.Run("VerifyNameConstraintsWithSAN", func(t *testing.T) {
-		caCert, err := x509.ParseCertificate(caCertBytes)
-		require.NoError(t, err)
-
-		leafTemplate := leafTemplate
-		leafTemplate.DNSNames = []string{"localhost"}
-
-		leafCertBytes, err := x509.CreateCertificate(rand.Reader, &leafTemplate, caCert, leafKey.Public(), caKey)
-		require.NoError(t, err)
-
-		leafCert, err := x509.ParseCertificate(leafCertBytes)
-		require.NoError(t, err)
-
-		err = verifyLegacyNameConstraints([]*x509.Certificate{leafCert, caCert})
-		require.NoError(t, err, "signer with name constraints and leaf with SANs should be valid")
-	})
-
-	t.Run("ValidationAtSetup", func(t *testing.T) {
-		fabricMSPConfig := &msp.FabricMSPConfig{
-			Name:      "ConstraintsMSP",
-			RootCerts: [][]byte{caCertPem},
-			SigningIdentity: &msp.SigningIdentityInfo{
-				PublicSigner: leafCertPem,
-				PrivateSigner: &msp.KeyInfo{
-					KeyIdentifier: "Certificate Without SAN",
-					KeyMaterial:   leafKeyPem,
-				},
-			},
-		}
-		mspConfig := &msp.MSPConfig{
-			Config: protoutil.MarshalOrPanic(fabricMSPConfig),
-		}
-
-		ks, err := sw.NewFileBasedKeyStore(nil, filepath.Join(configtest.GetDevMspDir(), "keystore"), true)
-		require.NoError(t, err)
-		cryptoProvider, err := sw.NewDefaultSecurityLevelWithKeystore(ks)
-		require.NoError(t, err)
-
-		testMSP, err := NewBccspMspWithKeyStore(MSPv1_0, ks, cryptoProvider)
-		require.NoError(t, err)
-
-		err = testMSP.Setup(mspConfig)
-		require.Error(t, err)
-		var cie x509.CertificateInvalidError
-		require.True(t, errors.As(err, &cie))
-		require.Equal(t, x509.NameConstraintsWithoutSANs, cie.Reason)
-	})
-}
-
 func TestIsWellFormed(t *testing.T) {
 	mspMgr := NewMSPManager()
 
@@ -524,7 +367,7 @@ func TestIsWellFormed(t *testing.T) {
 		T: big.NewInt(100),
 	})
 	require.NoError(t, err)
-	newCert, err := certFromX509Cert(cert)
+	newCert, err := bccsp.CertFromX509Cert(cert)
 	require.NoError(t, err)
 	newCert.SignatureValue.Bytes = modifiedSig
 	newCert.SignatureValue.BitLength = len(newCert.SignatureValue.Bytes) * 8
@@ -822,8 +665,10 @@ func TestGetOUFail(t *testing.T) {
 func TestCertificationIdentifierComputation(t *testing.T) {
 	id, err := localMsp.GetDefaultSigningIdentity()
 	require.NoError(t, err)
+	if id == nil {
 
-	chain, err := localMsp.(*bccspmsp).getCertificationChainForBCCSPIdentity(id.GetPublicVersion().(*identity))
+	}
+	chain, err := bccsp.GetValidationChain(localMsp.(*bccspmsp).opts, localMsp.(*bccspmsp).certificationTreeInternalNodesMap, id.GetPublicVersion().(*identity).cert, false)
 	require.NoError(t, err)
 
 	// Hash the chain
@@ -1606,4 +1451,64 @@ func TestProviderTypeToString(t *testing.T) {
 	// Check that the provider type is not found
 	pt = ProviderTypeToString(OTHER)
 	require.Equal(t, "", pt)
+}
+
+func generateSelfSignedCert(t *testing.T, now time.Time) (*ecdsa.PrivateKey, *x509.Certificate) {
+	k, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	require.NoError(t, err)
+
+	// Generate a self-signed certificate
+	testExtKeyUsage := []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth}
+	testUnknownExtKeyUsage := []asn1.ObjectIdentifier{[]int{1, 2, 3}, []int{2, 59, 1}}
+	extraExtensionData := []byte("extra extension")
+	commonName := "test.example.com"
+	template := x509.Certificate{
+		SerialNumber: big.NewInt(1),
+		Subject: pkix.Name{
+			CommonName:   commonName,
+			Organization: []string{"Σ Acme Co"},
+			Country:      []string{"US"},
+			ExtraNames: []pkix.AttributeTypeAndValue{
+				{
+					Type:  []int{2, 5, 4, 42},
+					Value: "Gopher",
+				},
+				// This should override the Country, above.
+				{
+					Type:  []int{2, 5, 4, 6},
+					Value: "NL",
+				},
+			},
+		},
+		NotBefore:             now.Add(-1 * time.Hour),
+		NotAfter:              now.Add(1 * time.Hour),
+		SignatureAlgorithm:    x509.ECDSAWithSHA256,
+		SubjectKeyId:          []byte{1, 2, 3, 4},
+		KeyUsage:              x509.KeyUsageCertSign,
+		ExtKeyUsage:           testExtKeyUsage,
+		UnknownExtKeyUsage:    testUnknownExtKeyUsage,
+		BasicConstraintsValid: true,
+		IsCA:                  true,
+		OCSPServer:            []string{"http://ocurrentBCCSP.example.com"},
+		IssuingCertificateURL: []string{"http://crt.example.com/ca1.crt"},
+		DNSNames:              []string{"test.example.com"},
+		EmailAddresses:        []string{"gopher@golang.org"},
+		IPAddresses:           []net.IP{net.IPv4(127, 0, 0, 1).To4(), net.ParseIP("2001:4860:0:2001::68")},
+		PolicyIdentifiers:     []asn1.ObjectIdentifier{[]int{1, 2, 3}},
+		PermittedDNSDomains:   []string{".example.com", "example.com"},
+		CRLDistributionPoints: []string{"http://crl1.example.com/ca1.crl", "http://crl2.example.com/ca1.crl"},
+		ExtraExtensions: []pkix.Extension{
+			{
+				Id:    []int{1, 2, 3, 4},
+				Value: extraExtensionData,
+			},
+		},
+	}
+	certRaw, err := x509.CreateCertificate(rand.Reader, &template, &template, &k.PublicKey, k)
+	require.NoError(t, err)
+
+	cert, err := x509.ParseCertificate(certRaw)
+	require.NoError(t, err)
+
+	return k, cert
 }
